+++ test/QuadPassport/test_set_attributes.ts
@@ -11,11 +11,15 @@ const {
   ATTRIBUTE_IS_BUSINESS,
   ATTRIBUTE_COUNTRY,
   READER_ROLE,
+  TOKEN_ID,
+  HARDHAT_CHAIN_ID,
 } = require("../../utils/constant.ts");
 
 const {
   deployPassportEcosystem,
 } = require("../helpers/deployment_and_init.ts");
+
+const { signSetAttributes, signAccount } = require("../helpers/signature.ts");
 
 const { setAttributes } = require("../helpers/set_attributes.ts");
 const { assertSetAttribute } = require("../helpers/asserts.ts");
@@ -36,7 +40,7 @@ describe("QuadPassport.setAttributes", async () => {
     issuerTreasury2: SignerWithAddress,
     mockReader: SignerWithAddress;
 
-  let issuedAt: number;
+  let issuedAt: number, verifiedAt: number;
   const attributes: any = {
     [ATTRIBUTE_DID]: formatBytes32String("did:quad:123456789abcdefghi"),
     [ATTRIBUTE_AML]: formatBytes32String("1"),
@@ -59,3069 +63,573 @@ describe("QuadPassport.setAttributes", async () => {
     ] = await ethers.getSigners();
     [governance, passport, reader, defi] = await deployPassportEcosystem(
       admin,
-      [issuer],
+      [issuer, issuer2],
       treasury,
-      [issuerTreasury]
+      [issuerTreasury, issuerTreasury2]
     );
 
     issuedAt = Math.floor(new Date().getTime() / 1000) - 100;
+    verifiedAt = Math.floor(new Date().getTime() / 1000) - 100;
+
+    await governance.connect(admin).grantRole(READER_ROLE, mockReader.address);
   });
 
-  describe("setAttribute", async () => {
-    it("success - setAttribute(AML)", async () => {});
+  describe("QuadPassport.setAttributes (success)", async () => {
+    beforeEach(async () => {});
 
-    it("success - setAttribute(IS_BUSINESS) EOA", async () => {
-      var newIsBusiness = id("TRUE");
-      var newIssuedAt = 1010;
-      await assertSetAttribute(
+    it("setAttributes (Single Attribute)", async () => {
+      const attributes: any = {
+        [ATTRIBUTE_IS_BUSINESS]: id("FALSE"),
+      };
+      await setAttributes(
         minterA,
         issuer,
-        issuerTreasury,
         passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-
-      newIsBusiness = id("FALSE");
-      newIssuedAt = 1011;
-      await assertSetAttribute(
-        minterA,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - setAttribute(IS_BUSINESS) Smart Contract", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newIsBusiness = id("FALSE");
-      await assertSetAttribute(
-        mockBusiness,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
+        attributes,
+        verifiedAt,
         issuedAt,
-        { signer: minterA }
+        MINT_PRICE
       );
-      await assertGetAttributeFree(
-        [issuer.address],
-        mockBusiness,
-        defi,
+      await assertSetAttribute(
+        minterA,
+        [issuer],
         passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
+        [attributes],
+        [verifiedAt],
+        [MINT_PRICE],
+        mockReader
+      );
+    });
+
+    it("setAttributes (Multiple Attribute)", async () => {
+      await setAttributes(
+        minterA,
+        issuer,
+        passport,
+        attributes,
+        verifiedAt,
         issuedAt,
-        1,
-        {
-          signer: minterA,
-          isContract: true,
-        }
+        MINT_PRICE
       );
 
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
+      await assertSetAttribute(
+        minterA,
+        [issuer],
         passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
+        [attributes],
+        [verifiedAt],
+        [MINT_PRICE],
+        mockReader
+      );
+    });
+
+    it("setAttributes (Multiple issuers for exact same Attribute)", async () => {
+      await setAttributes(
+        minterA,
+        issuer,
+        passport,
+        attributes,
+        verifiedAt,
         issuedAt,
-        1,
-        { signer: minterA }
+        MINT_PRICE
       );
-      // OK Fetching old value
-      await assertGetAttribute(
+
+      await setAttributes(
         minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
+        issuer2,
         passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
+        attributes,
+        verifiedAt,
         issuedAt,
-        1,
-        { signer: minterA }
+        MINT_PRICE
+      );
+
+      await assertSetAttribute(
+        minterA,
+        [issuer, issuer2],
+        passport,
+        [attributes, attributes],
+        [verifiedAt, verifiedAt],
+        [MINT_PRICE, MINT_PRICE],
+        mockReader
       );
     });
 
-    it("success - Individual setAttribute(COUNTRY)", async () => {
-      const newCountry = id("DE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await assertSetAttribute(
+    it("setAttributes (Multiple Issuers - same wallet same DID)", async () => {
+      await setAttributes(
         minterA,
         issuer,
-        issuerTreasury,
         passport,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
+        attributes,
+        verifiedAt,
+        issuedAt,
+        MINT_PRICE
       );
 
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
+      const attributeByIssuer2 = {
+        [ATTRIBUTE_DID]: formatBytes32String("did:quad:123456789abcdefghi"),
+        [ATTRIBUTE_AML]: formatBytes32String("9"),
+        [ATTRIBUTE_COUNTRY]: id("US"),
+        [ATTRIBUTE_IS_BUSINESS]: id("FALSE"),
+      };
+
+      await governance.connect(admin).setEligibleTokenId(2, true);
+
+      await setAttributes(
         minterA,
-        defi,
+        issuer2,
         passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
+        attributeByIssuer2,
+        verifiedAt + 1,
+        issuedAt + 1,
+        MINT_PRICE.add(1),
+        2 // new TokenId
       );
-      // OK Fetching old value
-      await assertGetAttribute(
+      await assertSetAttribute(
         minterA,
-        treasury,
+        [issuer, issuer2],
+        passport,
+        [attributes, attributeByIssuer2],
+        [verifiedAt, verifiedAt + 1],
+        [MINT_PRICE, MINT_PRICE.add(1)],
+        mockReader
+      );
+
+      expect(await passport.balanceOf(minterA.address, 1)).to.equal(1);
+      expect(await passport.balanceOf(minterA.address, 2)).to.equal(1);
+    });
+
+    it("setAttributes (Multiple Issuers - same wallet but diff DID)", async () => {
+      await setAttributes(
+        minterA,
         issuer,
-        issuerTreasury,
-        usdc,
-        defi,
         passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
+        attributes,
+        verifiedAt,
+        issuedAt,
+        MINT_PRICE
       );
-      const expectedPayment = PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY];
+
+      const attributeByIssuer2 = {
+        [ATTRIBUTE_DID]: formatBytes32String("did:quad:newdid"),
+        [ATTRIBUTE_AML]: formatBytes32String("9"),
+        [ATTRIBUTE_COUNTRY]: id("US"),
+        [ATTRIBUTE_IS_BUSINESS]: id("FALSE"),
+      };
+
+      await setAttributes(
+        minterA,
+        issuer2,
+        passport,
+        attributeByIssuer2,
+        verifiedAt + 1,
+        issuedAt + 1,
+        MINT_PRICE.add(1)
+      );
+
+      // Check DID
+      let response = await passport
+        .connect(mockReader)
+        .attributes(minterA.address, ATTRIBUTE_DID);
+
+      expect(response.length).equals(2);
+
+      expect(response[0].value).equals(attributes[ATTRIBUTE_DID]);
+      expect(response[0].issuer).equals(issuer.address);
+      expect(response[0].epoch).equals(verifiedAt);
+
+      expect(response[1].value).equals(attributeByIssuer2[ATTRIBUTE_DID]);
+      expect(response[1].issuer).equals(issuer2.address);
+      expect(response[1].epoch).equals(verifiedAt + 1);
+
+      // Check AML
+      response = await passport
+        .connect(mockReader)
+        .attributes(minterA.address, ATTRIBUTE_AML);
+
+      expect(response.length).equals(1);
+
+      expect(response[0].value).equals(attributes[ATTRIBUTE_AML]);
+      expect(response[0].issuer).equals(issuer.address);
+      expect(response[0].epoch).equals(verifiedAt);
+
+      // Check IS_BUSINESS
+      response = await passport
+        .connect(mockReader)
+        .attributes(minterA.address, ATTRIBUTE_IS_BUSINESS);
+
+      expect(response.length).equals(2);
+
+      expect(response[0].value).equals(
+        attributeByIssuer2[ATTRIBUTE_IS_BUSINESS]
+      );
+      expect(response[0].value).equals(response[1].value);
+      expect(response[0].issuer).equals(issuer.address);
+      expect(response[1].issuer).equals(issuer2.address);
+      expect(response[0].epoch).equals(verifiedAt);
+      expect(response[1].epoch).equals(verifiedAt + 1);
+
+      expect(await passport.balanceOf(minterA.address, TOKEN_ID)).to.equal(1);
       expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance.add(expectedPayment)
+        MINT_PRICE.mul(2).add(1)
       );
     });
 
-    it("success - Business setAttribute(COUNTRY)", async () => {
-      const sig = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sig,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newCountry = id("DE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-
-      await assertSetAttribute(
-        minterB,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-      const expectedPayment = PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY];
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance.add(expectedPayment)
-      );
-    });
-
-    it("success - Individuals may be assiged new eligible attributes", async () => {
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(false);
-      await governance.connect(admin).setEligibleAttribute(id("GENDER"), true);
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(true);
-
-      const sig = await signSetAttribute(
-        issuer,
+    it("setAttributes (fee = 0)", async () => {
+      const attributes: any = {
+        [ATTRIBUTE_IS_BUSINESS]: id("FALSE"),
+      };
+      await setAttributes(
         minterA,
-        TOKEN_ID,
-        id("GENDER"),
-        id("M"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .setAttribute(
-          minterA.address,
-          TOKEN_ID,
-          id("GENDER"),
-          id("M"),
-          issuedAt,
-          sig,
-          {
-            value: 0,
-          }
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        id("GENDER"),
-        id("M"),
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
         issuer,
-        issuerTreasury,
-        usdc,
-        defi,
         passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
+        attributes,
+        verifiedAt,
+        issuedAt,
+        0
       );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - Businesses may be assiged new eligible attributes", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(false);
-      await governance.connect(admin).setEligibleAttribute(id("GENDER"), true);
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(true);
-
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        id("GENDER"),
-        id("M"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterB)
-        .setAttribute(
-          minterB.address,
-          TOKEN_ID,
-          id("GENDER"),
-          id("M"),
-          issuedAt,
-          sig,
-          {
-            value: 0,
-          }
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        id("GENDER"),
-        id("M"),
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - Individual setAttribute(AML)", async () => {
-      const newAML = hexZeroPad("0x05", 32);
-      const newIssuedAt = 1000;
-      const initialBalance = await ethers.provider.getBalance(passport.address);
       await assertSetAttribute(
         minterA,
-        issuer,
-        issuerTreasury,
+        [issuer],
         passport,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
+        [attributes],
+        [verifiedAt],
+        [0],
+        mockReader
       );
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-      const expectedPayment = PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY];
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance.add(expectedPayment)
-      );
-    });
-
-    it("success - Business setAttribute(AML)", async () => {
-      const sig = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sig,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newAML = hexZeroPad("0x05", 32);
-      const newIssuedAt = 1000;
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await assertSetAttribute(
-        minterB,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-      const expectedPayment = PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY];
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance.add(expectedPayment)
-      );
-    });
-
-    it("success - mintPricePerAttribute(0)", async () => {
-      await passport.withdraw(issuerTreasury.address);
-      await governance
-        .connect(admin)
-        .setAttributeMintPrice(ATTRIBUTE_AML, parseEther("0"));
-
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await passport
-        .connect(minterA)
-        .setAttribute(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_AML,
-          newAML,
-          newIssuedAt,
-          sig,
-          {
-            value: parseEther("0"),
-          }
-        );
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(0);
-
-      await expect(
-        passport.callStatic.withdraw(issuerTreasury.address)
-      ).to.revertedWith("NOT_ENOUGH_BALANCE");
-    });
-
-    it("success - two issuers", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await assertSetAttribute(
-        minterA,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-
-      await passport.withdraw(issuerTreasury.address);
-
-      // Update AML with Issuer 2
-      const sig = await signMint(
-        issuerB,
-        minterA,
-        1,
-        id("DID_34"),
-        aml,
-        country,
-        isBusiness,
-        issuedAt
-      );
-
-      const sigAccount = await signMessage(minterA, minterA.address);
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            minterA.address,
-            1,
-            id("DID_34"),
-            aml,
-            country,
-            isBusiness,
-            issuedAt,
-          ],
-          sig,
-          sigAccount,
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newAML = id("MEDIUM");
-      await assertSetAttribute(
-        minterA,
-        issuerB,
-        issuerBTreasury,
-        passport,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-
-      // minterA now has attributes from both issuers
-      await assertGetAttributeIncluding(
-        minterA,
-        treasury,
-        [issuer.address, issuerB.address],
-        usdc,
-        defi,
-        governance,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        [newCountry, country],
-        [BigNumber.from(newIssuedAt), BigNumber.from(issuedAt)],
-        [issuer.address, issuerB.address],
-        1,
-        {}
-      );
-    });
-
-    it("succes - change issuer treasury", async () => {
-      await passport.withdraw(issuerTreasury.address);
-      const newIssuerTreasury = ethers.Wallet.createRandom();
-      await governance
-        .connect(admin)
-        .setIssuer(issuer.address, newIssuerTreasury.address);
-
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await assertSetAttribute(
-        minterA,
-        issuer,
-        newIssuerTreasury,
-        passport,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        newIssuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-
-      await expect(passport.withdraw(issuerTreasury.address)).to.revertedWith(
-        "NOT_ENOUGH_BALANCE"
-      );
-    });
-
-    it("fail - setAttribute(DID) as Individual", async () => {
-      const newDid = formatBytes32String("did:1:newdid");
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_DID,
-        newDid,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_DID,
-            newDid,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_DID],
-            }
-          )
-      ).to.revertedWith("MUST_BURN_AND_MINT");
-    });
-
-    it("fail - setAttribute(DID) as Business", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newDid = formatBytes32String("did:1:newdid");
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_DID,
-        newDid,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_DID,
-            newDid,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_DID],
-            }
-          )
-      ).to.revertedWith("MUST_BURN_AND_MINT");
-    });
-
-    it("fail - passport tokenId invalid (Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const invalidTokenId = 2;
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        invalidTokenId,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            invalidTokenId,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt,
-            sig,
-            { value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY] }
-          )
-      ).to.revertedWith("PASSPORT_TOKENID_INVALID");
-    });
-
-    it("fail - passport tokenId invalid (Individual)", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const invalidTokenId = 2;
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        invalidTokenId,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            invalidTokenId,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt,
-            sig,
-            { value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY] }
-          )
-      ).to.revertedWith("PASSPORT_TOKENID_INVALID");
-    });
-
-    it("fail - no passport", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt,
-            sig,
-            { value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY] }
-          )
-      ).to.revertedWith("PASSPORT_DOES_NOT_EXIST");
-    });
-
-    it("fail - attribute not eligible (Individual)", async () => {
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        true
-      );
-      await governance
-        .connect(admin)
-        .setEligibleAttribute(ATTRIBUTE_COUNTRY, false);
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        false
-      );
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY],
-            }
-          )
-      ).to.revertedWith("ATTRIBUTE_NOT_ELIGIBLE");
-    });
-
-    it("fail - attribute not eligible (Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        true
-      );
-      await governance
-        .connect(admin)
-        .setEligibleAttribute(ATTRIBUTE_COUNTRY, false);
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        false
-      );
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY],
-            }
-          )
-      ).to.revertedWith("ATTRIBUTE_NOT_ELIGIBLE");
-    });
-
-    it("fail - invalid mint price too high per attribute (Individual)", async () => {
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: parseEther("1"),
-            }
-          )
-      ).to.revertedWith("INVALID_ATTR_MINT_PRICE");
-    });
-
-    it("fail - invalid mint too high price per attribute (Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: parseEther("1"),
-            }
-          )
-      ).to.revertedWith("INVALID_ATTR_MINT_PRICE");
-    });
-
-    it("fail - invalid mint price too low per attribute (Individual)", async () => {
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: parseEther("0.0000001"),
-            }
-          )
-      ).to.revertedWith("INVALID_ATTR_MINT_PRICE");
-    });
-
-    it("fail - invalid mint too low price per attribute (Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            country,
-            issuedAt,
-            sig,
-            {
-              value: parseEther("0.0000001"),
-            }
-          )
-      ).to.revertedWith("INVALID_ATTR_MINT_PRICE");
-    });
-
-    it("fail - signature already used for AML", async () => {
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await assertSetAttribute(
-        minterA,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            newAML,
-            newIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("SIGNATURE_ALREADY_USED");
-    });
-
-    it("fail - signature already used for IS_BUSINESS as Individual", async () => {
-      const newIsBusiness = id("TRUE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await assertSetAttribute(
-        minterA,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_IS_BUSINESS,
-            newIsBusiness,
-            newIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_IS_BUSINESS],
-            }
-          )
-      ).to.revertedWith("SIGNATURE_ALREADY_USED");
-    });
-
-    it("fail - signature already used for IS_BUSINESS as Business", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newIsBusiness = id("TRUE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await assertSetAttribute(
-        minterB,
-        issuer,
-        issuerTreasury,
-        passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_IS_BUSINESS,
-            newIsBusiness,
-            newIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_IS_BUSINESS],
-            }
-          )
-      ).to.revertedWith("SIGNATURE_ALREADY_USED");
-    });
-
-    it("success - anyone may sign for minterB", async () => {
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await passport
-        .connect(minterB)
-        .setAttribute(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_AML,
-          newAML,
-          newIssuedAt,
-          sig,
-          {
-            value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-          }
-        );
-    });
-
-    it("fail - not issuer role (Individual)", async () => {
-      const newIssuer = ethers.Wallet.createRandom();
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const sig = await signSetAttribute(
-        newIssuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            newAML,
-            newIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - not issuer role (Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newIssuer = ethers.Wallet.createRandom();
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const sig = await signSetAttribute(
-        newIssuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            newAML,
-            newIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (tokenId)", async () => {
-      const wrongTokenId = 2;
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            wrongTokenId,
-            ATTRIBUTE_AML,
-            aml,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (attribute type, Individual)", async () => {
-      const wrongAttribute = ATTRIBUTE_COUNTRY;
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            wrongAttribute,
-            aml,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (attribute type, Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const wrongAttribute = ATTRIBUTE_COUNTRY;
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            wrongAttribute,
-            aml,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (attribute value for AML)", async () => {
-      const wrongAML = id("HIGH");
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            wrongAML,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (attribute value for COUNTRY, Inidividual)", async () => {
-      const wrongAML = id("FR");
-      aml = id("DE");
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            wrongAML,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (attribute value for COUNTRY, Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const wrongAML = id("FR");
-      aml = id("DE");
-
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_COUNTRY,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            wrongAML,
-            issuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_COUNTRY],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (issuedAt, Individual)", async () => {
-      const wrongIssuedAt = issuedAt + 1;
-      const sig = await signSetAttribute(
-        issuer,
-        minterA,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterA)
-          .setAttribute(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            aml,
-            wrongIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - invalid sig (issuedAt, Business)", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const wrongIssuedAt = issuedAt + 1;
-      const sig = await signSetAttribute(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      await expect(
-        passport
-          .connect(minterB)
-          .setAttribute(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_AML,
-            aml,
-            wrongIssuedAt,
-            sig,
-            {
-              value: PRICE_SET_ATTRIBUTE[ATTRIBUTE_AML],
-            }
-          )
-      ).to.revertedWith("INVALID_ISSUER");
     });
   });
 
-  describe("setAttributeIssuer", async () => {
-    it("success - setAttributeIssuer(AML)", async () => {
-      const newAML = id("HIGH");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_AML,
-          newAML,
-          newIssuedAt
-        );
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        newIssuedAt
-      );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-    });
+  describe("QuadPassport.setAttributes (ALL ERRORS)", async () => {
+    let attrKeys: any,
+      attrValues: any,
+      fee: any,
+      tokenId: any,
+      blockId: any,
+      sigIssuer: any,
+      sigAccount: any;
 
-    it("success - Individual setAttribute(AML=5)", async () => {
-      await passport.withdraw(issuerTreasury.address);
+    beforeEach(async () => {
+      attrKeys = [];
+      attrValues = [];
 
-      const newAML = hexZeroPad("0x05", 32);
-      const newIssuedAt = 1000;
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_AML,
-          newAML,
-          newIssuedAt
-        );
-
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-      await expect(passport.withdraw(issuerTreasury.address)).to.revertedWith(
-        "NOT_ENOUGH_BALANCE"
-      );
-    });
-
-    it("success - Business setAttribute(AML=5)", async () => {
-      const sig = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sig,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      await passport.withdraw(issuerTreasury.address);
-
-      const newAML = hexZeroPad("0x05", 32);
-      const newIssuedAt = 1000;
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterB.address,
-          TOKEN_ID,
-          ATTRIBUTE_AML,
-          newAML,
-          newIssuedAt
-        );
-
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        newAML,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-      await expect(passport.withdraw(issuerTreasury.address)).to.revertedWith(
-        "NOT_ENOUGH_BALANCE"
-      );
-    });
-
-    it("success - Individual setAttribute(COUNTRY)", async () => {
-      await passport.withdraw(issuerTreasury.address);
-      const newCountry = id("DE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_COUNTRY,
-          newCountry,
-          newIssuedAt
-        );
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-      await expect(passport.withdraw(issuerTreasury.address)).to.revertedWith(
-        "NOT_ENOUGH_BALANCE"
-      );
-    });
-
-    it("success - Business setAttribute(COUNTRY)", async () => {
-      const sig = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sig,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      await passport.withdraw(issuerTreasury.address);
-      const newCountry = id("DE");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterB.address,
-          TOKEN_ID,
-          ATTRIBUTE_COUNTRY,
-          newCountry,
-          newIssuedAt
-        );
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-      await expect(passport.withdraw(issuerTreasury.address)).to.revertedWith(
-        "NOT_ENOUGH_BALANCE"
-      );
-    });
-    it("success - setAttribute(IS_BUSINESS) Business", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      var newIsBusiness = id("FALSE");
-
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          mockBusiness.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          issuedAt
-        );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        mockBusiness,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        issuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
+      Object.keys(attributes).forEach((k, i) => {
+        let attrKey;
+        if (k === ATTRIBUTE_AML) {
+          attrKey = ethers.utils.keccak256(
+            ethers.utils.defaultAbiCoder.encode(
+              ["bytes32", "bytes32"],
+              [attributes[ATTRIBUTE_DID], k]
+            )
+          );
+        } else {
+          attrKey = ethers.utils.keccak256(
+            ethers.utils.defaultAbiCoder.encode(
+              ["address", "bytes32"],
+              [minterA.address, k]
+            )
+          );
         }
-      );
+        attrKeys.push(attrKey);
+        attrValues.push(attributes[k]);
+      });
 
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
+      fee = MINT_PRICE;
+      tokenId = TOKEN_ID;
+      blockId = HARDHAT_CHAIN_ID;
+
+      sigIssuer = await signSetAttributes(
+        minterA,
         issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
+        attributes,
+        verifiedAt,
         issuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
-        }
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
-        }
+        fee,
+        blockId,
+        tokenId
       );
 
-      newIsBusiness = id("TRUE");
-      const newIssuedAt = 1011;
-
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          mockBusiness.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          newIssuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        mockBusiness,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
-        }
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
-        }
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt,
-        1,
-        {
-          signer: minterA,
-          mockBusiness: mockBusiness,
-        }
-      );
+      sigAccount = await signAccount(minterA);
     });
 
-    it("success - setAttribute(IS_BUSINESS) Individual", async () => {
-      var newIsBusiness = id("TRUE");
-      var newIssuedAt = 1010;
-      const initialBalance = await ethers.provider.getBalance(passport.address);
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          newIssuedAt
-        );
-      expect(await ethers.provider.getBalance(passport.address)).to.equal(
-        initialBalance
-      );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-
-      newIsBusiness = id("FALSE");
-      newIssuedAt = 1011;
-      await assertSetAttribute(
+    it("fail - signature already used", async () => {
+      await setAttributes(
         minterA,
         issuer,
-        issuerTreasury,
         passport,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - Individuals may be assiged new eligible attributes", async () => {
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(false);
-      await governance.connect(admin).setEligibleAttribute(id("GENDER"), true);
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(true);
-
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          id("GENDER"),
-          id("F"),
-          issuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        id("GENDER"),
-        id("F"),
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - Businesses may be assiged new eligible attributes", async () => {
-      const sigMint = await signMint(
-        issuer,
-        minterB,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [minterB.address, TOKEN_ID, did, aml, country, id("TRUE"), issuedAt],
-          sigMint,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(false);
-      await governance.connect(admin).setEligibleAttribute(id("GENDER"), true);
-      expect(await governance.eligibleAttributes(id("GENDER"))).to.equal(true);
-
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterB.address,
-          TOKEN_ID,
-          id("GENDER"),
-          id("F"),
-          issuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        id("GENDER"),
-        id("F"),
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterB,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_AML,
-        aml,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterB,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - setAttribute(IS_BUSINESS) Individual", async () => {
-      var newIsBusiness = id("TRUE");
-      var newIssuedAt = 1010;
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          newIssuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-
-      newIsBusiness = id("FALSE");
-      newIssuedAt = 1011;
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          newIssuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        minterA,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
-        newIssuedAt
-      );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt
-      );
-    });
-
-    it("success - setAttribute(IS_BUSINESS) Smart Contract", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newIsBusiness = id("FALSE");
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          mockBusiness.address,
-          TOKEN_ID,
-          ATTRIBUTE_IS_BUSINESS,
-          newIsBusiness,
-          issuedAt
-        );
-
-      await assertGetAttributeFree(
-        [issuer.address],
-        mockBusiness,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_IS_BUSINESS,
-        newIsBusiness,
+        attributes,
+        verifiedAt,
         issuedAt,
-        1,
-        {
-          signer: minterA,
-          isContract: true,
-        }
+        MINT_PRICE
       );
-
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        country,
-        issuedAt,
-        1,
-        { signer: minterA }
-      );
-      // OK Fetching old value
-      await assertGetAttribute(
-        mockBusiness,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_DID,
-        did,
-        issuedAt,
-        1,
-        { signer: minterA }
-      );
-    });
-
-    it("success - two issuers (issuers may not overwrite each other)", async () => {
-      let newCountry = id("USA");
-      let newIssuedAt = Math.floor(new Date().getTime() / 1000);
-
-      // Update Country with Issuer 1
-      await passport
-        .connect(issuer)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_COUNTRY,
-          newCountry,
-          newIssuedAt
-        );
-      await assertGetAttribute(
-        minterA,
-        treasury,
-        issuer,
-        issuerTreasury,
-        usdc,
-        defi,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        newCountry,
-        newIssuedAt
-      );
-
-      const newIssuedAt2 = Math.floor(new Date().getTime() / 1000) + 100;
-
-      // Update Country with Issuer 2
-      await passport
-        .connect(issuerB)
-        .setAttributeIssuer(
-          minterA.address,
-          TOKEN_ID,
-          ATTRIBUTE_COUNTRY,
-          id("FRANCE"),
-          newIssuedAt2
-        );
-
-      await passport.withdrawToken(issuerTreasury.address, usdc.address);
-
-      // minterA now has attributes from both issuers
-      await assertGetAttributeIncluding(
-        minterA,
-        treasury,
-        [issuer.address, issuerB.address],
-        usdc,
-        defi,
-        governance,
-        passport,
-        reader,
-        ATTRIBUTE_COUNTRY,
-        [newCountry, id("FRANCE")],
-        [BigNumber.from(newIssuedAt), BigNumber.from(newIssuedAt2)],
-        [issuer.address, issuerB.address],
-        1,
-        {}
-      );
-    });
-
-    it("fail - zero address", async () => {
-      const newCountry = id("USA");
-
       await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            "0x0000000000000000000000000000000000000000",
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            issuedAt
-          )
-      ).to.revertedWith("ACCOUNT_CANNOT_BE_ZERO");
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          verifiedAt,
+          issuedAt,
+          MINT_PRICE
+        )
+      ).to.be.revertedWith("SIGNATURE_ALREADY_USED");
+    });
+
+    it("fail - invalid tokenId", async () => {
+      const badTokenId = 1337;
+      await expect(
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          verifiedAt,
+          issuedAt,
+          MINT_PRICE,
+          badTokenId
+        )
+      ).to.be.revertedWith("PASSPORT_TOKENID_INVALID");
+    });
+
+    it("fail - zero verifiedAt", async () => {
+      const verifiedAt = 0;
+      await expect(
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          verifiedAt,
+          issuedAt,
+          MINT_PRICE
+        )
+      ).to.be.revertedWith("VERIFIED_AT_CANNOT_BE_ZERO");
+    });
+
+    it("fail - future verifiedAt", async () => {
+      const blockNumAfter = await ethers.provider.getBlockNumber();
+      const currentBlock = await ethers.provider.getBlock(blockNumAfter);
+      const badVerifiedAt = currentBlock.timestamp + 100;
+      await expect(
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          badVerifiedAt,
+          issuedAt,
+          MINT_PRICE
+        )
+      ).to.be.revertedWith("INVALID_VERIFIED_AT");
     });
 
     it("fail - zero issuedAt", async () => {
+      const badIssuedAt = 0;
       await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            id("USA"),
-            0
-          )
-      ).to.revertedWith("ISSUED_AT_CANNOT_BE_ZERO");
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          verifiedAt,
+          badIssuedAt,
+          MINT_PRICE
+        )
+      ).to.be.revertedWith("ISSUED_AT_CANNOT_BE_ZERO");
     });
 
-    it("fail - future issuedAt", async () => {
+    it("fail - issuedAt expired", async () => {
       const blockNumAfter = await ethers.provider.getBlockNumber();
-      const blockAfter = await ethers.provider.getBlock(blockNumAfter);
+      const currentBlock = await ethers.provider.getBlock(blockNumAfter);
+      const issuedAt = currentBlock.timestamp - 90400;
+      await expect(
+        setAttributes(
+          minterA,
+          issuer,
+          passport,
+          attributes,
+          verifiedAt,
+          issuedAt,
+          MINT_PRICE
+        )
+      ).to.be.revertedWith("EXPIRED_ISSUED_AT");
+    });
+
+    it("fail - invalid fee", async () => {
+      const wrongFee = fee.sub(1);
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: wrongFee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_SET_ATTRIBUTE_FEE");
+    });
+
+    it("fail - not an ISSUER_ROLE", async () => {
+      await governance.connect(admin).setIssuerStatus(issuer.address, false);
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_ISSUER");
+    });
+
+    it("fail - attrKeys.length != attrValues.length", async () => {
+      attrKeys.push(id("wrong"));
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("MISMATCH_LENGTH");
+    });
+
+    it("fail - invalid signature (attrKeys)", async () => {
+      attrKeys[0] = id("wrong");
 
       await expect(
         passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            id("USA"),
-            blockAfter.timestamp + 100
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
           )
-      ).to.revertedWith("INVALID_ISSUED_AT");
+      ).to.be.revertedWith("INVALID_ISSUER");
     });
 
-    it("fail - passport tokenId invalid as Individual", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
+    it("fail - invalid signature (attrValues)", async () => {
+      attrValues[0] = id("wrong");
+
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_ISSUER");
+    });
+
+    it("fail - invalid signature (chainId)", async () => {
+      const wrongChainId = 1;
+      sigIssuer = await signSetAttributes(
+        minterA,
+        issuer,
+        attributes,
+        verifiedAt,
+        issuedAt,
+        fee,
+        wrongChainId,
+        tokenId
+      );
+
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_ISSUER");
+    });
+
+    it("fail - invalid signature (issuedAt)", async () => {
+      const wrongIssuedAt = issuedAt - 1;
+
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, wrongIssuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_ISSUER");
+    });
+
+    it("fail - invalid signature (verifiedAt)", async () => {
+      const wrongVerifiedAt = verifiedAt - 1;
+
+      await expect(
+        passport
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, wrongVerifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
+          )
+      ).to.be.revertedWith("INVALID_ISSUER");
+    });
+
+    it("fail - invalid signature (tokenId)", async () => {
       const wrongTokenId = 2;
+      await governance.connect(admin).setEligibleTokenId(wrongTokenId, true);
 
       await expect(
         passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterA.address,
-            wrongTokenId,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, wrongTokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
           )
-      ).to.revertedWith("PASSPORT_TOKENID_INVALID");
+      ).to.be.revertedWith("INVALID_ISSUER");
     });
 
-    it("fail - passport tokenId invalid as Business", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      const wrongTokenId = 2;
-
+    it("fail - mismatch sigAccount and sigIssuer", async () => {
+      sigAccount = await signAccount(minterB);
       await expect(
         passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            mockBusiness.address,
-            wrongTokenId,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
+          .connect(minterA)
+          .setAttributes(
+            [attrKeys, attrValues, tokenId, verifiedAt, issuedAt, fee],
+            sigIssuer,
+            sigAccount,
+            {
+              value: fee,
+            }
           )
-      ).to.revertedWith("PASSPORT_TOKENID_INVALID");
-    });
-
-    it("fail - no passport", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-
-      await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterB.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
-          )
-      ).to.revertedWith("PASSPORT_DOES_NOT_EXIST");
-    });
-
-    it("fail - attribute not eligible as Individual", async () => {
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        true
-      );
-      await governance
-        .connect(admin)
-        .setEligibleAttribute(ATTRIBUTE_COUNTRY, false);
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        false
-      );
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-
-      await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
-          )
-      ).to.revertedWith("ATTRIBUTE_NOT_ELIGIBLE");
-    });
-
-    it("fail - attribute not eligible as Business", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        true
-      );
-      await governance
-        .connect(admin)
-        .setEligibleAttribute(ATTRIBUTE_COUNTRY, false);
-      expect(await governance.eligibleAttributes(ATTRIBUTE_COUNTRY)).to.equal(
-        false
-      );
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-
-      await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            mockBusiness.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
-          )
-      ).to.revertedWith("ATTRIBUTE_NOT_ELIGIBLE");
-    });
-
-    it("fail - invalid issuer role", async () => {
-      const newCountry = id("USA");
-      const newIssuedAt = Math.floor(new Date().getTime() / 1000);
-      await expect(
-        passport
-          .connect(admin)
-          .setAttributeIssuer(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_COUNTRY,
-            newCountry,
-            newIssuedAt
-          )
-      ).to.revertedWith("INVALID_ISSUER");
-    });
-
-    it("fail - setAttribute(DID) as Individual", async () => {
-      const newDid = formatBytes32String("did:1:newdid");
-      await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            minterA.address,
-            TOKEN_ID,
-            ATTRIBUTE_DID,
-            newDid,
-            issuedAt
-          )
-      ).to.revertedWith("MUST_BURN_AND_MINT");
-    });
-
-    it("fail - setAttribute(DID) as Business", async () => {
-      const MockBusiness = await ethers.getContractFactory("MockBusiness");
-      const mockBusiness = await MockBusiness.deploy(defi.address);
-      await mockBusiness.deployed();
-
-      const sigBusiness = await signMint(
-        issuer,
-        mockBusiness,
-        TOKEN_ID,
-        did,
-        aml,
-        country,
-        id("TRUE"),
-        issuedAt
-      );
-
-      await passport
-        .connect(minterA)
-        .mintPassport(
-          [
-            mockBusiness.address,
-            TOKEN_ID,
-            did,
-            aml,
-            country,
-            id("TRUE"),
-            issuedAt,
-          ],
-          sigBusiness,
-          "0x00",
-          {
-            value: MINT_PRICE,
-          }
-        );
-
-      const newDid = formatBytes32String("did:1:newdid");
-      await expect(
-        passport
-          .connect(issuer)
-          .setAttributeIssuer(
-            mockBusiness.address,
-            TOKEN_ID,
-            ATTRIBUTE_DID,
-            newDid,
-            issuedAt
-          )
-      ).to.revertedWith("MUST_BURN_AND_MINT");
+      ).to.be.revertedWith("INVALID_ISSUER");
     });
   });
 });
